Introduce type library embedder. This uses low level Win32 API to update a file's resources and therefore embed a generated type library into the source assembly, which allows distributing the assembly without having to distribute a type library alongside it.
Note that embedding a type library will not make sense for an AnyCPU assembly, particularly for net48 assemblies where the same assembly can be loaded by either 32-bit or 64-bit processes and shipping two separate type libraries are preferable. However for net48 x86/x64 or .net5 assemblies where comhost.dll are already tied to specific architecture, it can be helpful in simplifying the distribution.

This can be invoked by calling `--embedtlb` argument included to the tool.