fix: Repair ability to register an assembly after the build
Use 'dscom tlb(un)register' to register the TLBs after the build.
This is done in favor of 'dscom regasm (--unregister)', as the TLB has already been created before.

It is also possible, but not recommended to perform
a registration system-wide, but this must be
opted-in and the user will be warned about it.

Closes #367