fix: Use DsCom tlb(un)register after/before build to (un)register the resulting TLB (#372)
* qa: Add a test that performs registration after the build

By adding a test for dscom that performs a build using

DsComRegisterTypeLibrariesAfterBuild=true

We can test, that the registration of assemblies is working.

* fix: Repair ability to register an assembly after the build

Use 'dscom tlb(un)register' to register the TLBs after the build.
This is done in favor of 'dscom regasm (--unregister)', as the TLB has already been created before.

It is also possible, but not recommended to perform
a registration system-wide, but this must be
opted-in and the user will be warned about it.

Closes #367

* qa: Add another acceptance test with registration

As x86 dotnet SDK is not installed, the 32bit acceptance tests are not executed.
Hence, we need an acceptance test, that is actually
executed during pull requests.

* fix: Only unregister if registration has been performed previously

As there is no satisfying ability in msbuild to test
 if the TLB is registered previously, we will try to
 rely on a touch file.

---------

Co-authored-by: Carsten Igel <cigel@dspace.de>